server/requirements.txt
python-socketio
uvicorn